Move code to separate file, only available for swift 6
